Backtest/requirements.txt
backtrader>=1.9.78.123
pandas>=2.0
pyyaml>=6.0
pytest>=7.0
pyarrow>=13.0

